pipettes changes for SPIrefactor
