+++ common/firmware/spi/spi_comms.cpp
@@ -1,7 +1,7 @@
 #include "common/firmware/spi_comms.hpp"
 
-#include "spi.h"
 #include "platform_specific_hal_conf.h"
+#include "spi.h"
 
 using namespace spi;
 
@@ -26,17 +26,16 @@ using namespace spi;
  * Public Functions
  */
 
-Spi::Spi(SPI_interface SPI_intf_instance): SPI_intf(SPI_intf_instance) {
-}
+Spi::Spi(SPI_interface SPI_intf_instance) : SPI_intf(SPI_intf_instance) {}
 
 void Spi::transmit_receive(const BufferType& transmit, BufferType& receive) {
     Reset_CS_Pin(SPI_intf.GPIO_handle, SPI_intf.pin);
     // NOLINTNEXTLINE(cppcoreguidelines-pro-type-const-cast)
     hal_transmit_receive(const_cast<uint8_t*>(transmit.data()), receive.data(),
-                         BUFFER_SIZE, TIMEOUT, (SPI_HandleTypeDef *)this->SPI_intf.SPI_handle);
+                         BUFFER_SIZE, TIMEOUT,
+                         (SPI_HandleTypeDef*)this->SPI_intf.SPI_handle);
     /*
         pass GPIO_instance and pin
     */
     Set_CS_Pin(SPI_intf.GPIO_handle, SPI_intf.pin);
-    
 }
