+++ pipettes/firmware/can_task.cpp
@@ -22,6 +22,7 @@
 #include "motor-control/core/motor_messages.hpp"
 #include "pipettes/core/eeprom.hpp"
 #include "pipettes/core/message_handlers/eeprom.hpp"
+#include "platform_specific_hal_conf.h"
 
 using namespace hal_can_bus;
 using namespace can_messages;
@@ -47,7 +48,103 @@ static freertos_message_queue::FreeRTOSMessageQueue<Move> motor_queue(
 static freertos_message_queue::FreeRTOSMessageQueue<Ack> complete_queue(
     "Complete Queue");
 
-static spi::Spi spi_comms{};
+void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
+    GPIO_InitTypeDef GPIO_InitStruct = {0};
+    if (hspi->Instance == SPI2) {
+        /* Peripheral clock enable */
+        __HAL_RCC_SPI2_CLK_ENABLE();
+        __HAL_RCC_GPIOA_CLK_ENABLE();
+        __HAL_RCC_GPIOB_CLK_ENABLE();
+        /**SPI2 GPIO Configuration
+        PC6     ------> SPI2_CS
+        PB13     ------> SPI2_SCK
+        PB14     ------> SPI2_CIPO
+        PB15     ------> SPI2_COPI
+
+         Step/Dir
+         PC3  ---> Dir Pin
+         PC7  ---> Step Pin
+         Enable
+         PC8  ---> Enable Pin
+        */
+        GPIO_InitStruct.Pin = GPIO_PIN_13 | GPIO_PIN_14 | GPIO_PIN_15;
+        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
+        GPIO_InitStruct.Pull = GPIO_NOPULL;
+        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
+        GPIO_InitStruct.Alternate = GPIO_AF5_SPI2;
+        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
+
+        // Chip select
+        GPIO_InitStruct.Pin = GPIO_PIN_6;
+        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
+        HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
+
+        // Enable/Dir/Step pin
+        GPIO_InitStruct.Pin = GPIO_PIN_3 | GPIO_PIN_7 | GPIO_PIN_8;
+        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
+        GPIO_InitStruct.Pull = GPIO_NOPULL;
+        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
+        HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
+
+        GPIO_InitStruct.Pin = GPIO_PIN_7 | GPIO_PIN_8;
+        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
+        HAL_GPIO_Init(GPIOC,  // NOLINT(cppcoreguidelines-pro-type-cstyle-cast)
+                      &GPIO_InitStruct);
+    }
+}
+
+SPI_HandleTypeDef MX_SPI2_Init() {
+    /* SPI2 parameter configuration*/
+    __HAL_RCC_SPI2_CLK_ENABLE();
+    SPI_HandleTypeDef hspi2 = {
+        .Instance = SPI2,
+        .Init = {.Mode = SPI_MODE_MASTER,
+                 .Direction = SPI_DIRECTION_2LINES,
+                 .DataSize = SPI_DATASIZE_8BIT,
+                 .CLKPolarity = SPI_POLARITY_HIGH,
+                 .CLKPhase = SPI_PHASE_2EDGE,
+                 .NSS = SPI_NSS_SOFT,
+                 .BaudRatePrescaler = SPI_BAUDRATEPRESCALER_32,
+                 .FirstBit = SPI_FIRSTBIT_MSB,
+                 .TIMode = SPI_TIMODE_DISABLE,
+                 .CRCCalculation = SPI_CRCCALCULATION_DISABLE,
+                 .CRCPolynomial = 7,
+                 .CRCLength = SPI_CRC_LENGTH_DATASIZE,
+                 .NSSPMode = SPI_NSS_PULSE_DISABLE}
+
+    };
+
+    if (HAL_SPI_Init(&hspi2) != HAL_OK) {
+        Error_Handler();
+    }
+    return hspi2;
+}
+void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi) {
+    if (hspi->Instance == SPI2) {
+        /* Peripheral clock disable */
+        __HAL_RCC_SPI2_CLK_DISABLE();
+
+        /**SPI2 GPIO Configuration
+        PB12     ------> SPI2_NSS
+        PB13     ------> SPI2_SCK
+        PB14     ------> SPI2_MISO
+        PB15     ------> SPI2_MOSI
+        */
+        HAL_GPIO_DeInit(GPIOB,
+                        GPIO_PIN_12 | GPIO_PIN_13 | GPIO_PIN_14 | GPIO_PIN_15);
+        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_8 | GPIO_PIN_9);
+    }
+}
+
+SPI_HandleTypeDef handle = MX_SPI2_Init();
+
+spi::SPI_interface SPI_intf = {
+
+    .SPI_handle = &handle,
+    .GPIO_handle = GPIOB,
+    .pin = GPIO_PIN_12,
+};
+static spi::Spi spi_comms(SPI_intf);
 
 struct motion_controller::HardwareConfig PinConfigurations {
     .direction = {.port = GPIOC, .pin = GPIO_PIN_3},
